requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
PyPDF2==3.0.1
Pillow==10.0.0